\name{altitude}

\alias{altitude}

\docType{data}

\title{Virtual altitudinal data}

\description{Data frame with virtual altitudinal data. The data frame is
  used in the examples of the hSDM package vignette to derive an
  altitude raster determining species habitat suitability.}

\usage{altitude}

\format{
  
  \code{altitude} is a data frame with 2500 observations (50 x 50 cells) and 3 variables:
  \describe{
    \item{\code{x}}{coordinates of the center of the cell on the x axis}
    \item{\code{y}}{coordinates of the center of the cell on the y axis}
    \item{\code{altitude}}{altitude (m)}
  }
  
}

\keyword{datasets}