\name{cfr.env}

\alias{cfr.env}

\docType{data}

\title{Environmental data for South Africa's Cap Floristic Region}

\description{Data include environmental variables for 36909 one minute
by one minute grid cells on the whole South Africa's Cap Floristic
Region.}

\usage{cfr.env}

\format{
  
  \code{cfr.env} is a data frame with 36909 observations (cells) on the
  following six environmental variables.
  \describe{
    \item{\code{lon}}{longitude}
    \item{\code{lat}}{latitude}
    \item{\code{min07}}{minimum temperature of the coldest month (July)}
    \item{\code{smdwin}}{winter soil moisture days}
    \item{\code{fert3}}{moderately high fertility (percent of grid cell)}
    \item{\code{ph1}}{acidic soil (percent of grid cell)}
    \item{\code{text1}}{fine soil texture (percent of grid cell)}
    \item{\code{text2}}{moderately fine soil texture (percent of grid cell)}
  }
  
}

\source{Cory Merow's personal data}

\references{

  Latimer, A. M.; Wu, S. S.; Gelfand, A. E. and Silander, J. A. (2006)
  Building statistical models to analyze species
  distributions. \emph{Ecological Applications}, 16, 33-50.

}

\keyword{datasets}
