\name{data.Kery2010}

\alias{data.Kery2010}

\docType{data}

\title{Count data for the Willow tit (from Kéry and Royle 2010)}

\description{Repeated count data for the Willow tit (\emph{Poecile
  montanus}, a pesserine bird) in Switzerland on the period
  1999-2003. Data come from the Swiss national breeding bird survey MHB
  (Monitoring Haüfige Brutvögel).}

\usage{data.Kery2010}

\format{
  
  \code{data.Kery2010} is a data frame with 264 observations (1 km^2 quadrats) and the
  following 10 variables.
  \describe{
    \item{\code{coordx}}{quadrat x coordinate}
    \item{\code{coordy}}{quadrat y coordinate}
    \item{\code{elevation}}{mean quadrat elevation (m)}
    \item{\code{forest}}{quadrat forest cover (in \%)}
    \item{\code{count1}}{count for survey 1}
    \item{\code{count2}}{count for survey 2}
    \item{\code{count3}}{count for survey 3}
    \item{\code{juldate1}}{Julian date of survey 1}
    \item{\code{juldate2}}{Julian date of survey 2}
    \item{\code{juldate3}}{Julian date of survey 3}
  }
  
}

\source{Kéry and Royle, 2010, \emph{Journal of Animal Ecology}, 79, 453-461.}

\references{

  Kéry, M. and Andrew Royle, J. 2010. Hierarchical modelling and estimation of
  abundance and population trends in metapopulation designs. \emph{Journal of
  Animal Ecology}, 79, 453-461.

}

\keyword{datasets}
