\name{hSDM-package}
\alias{hSDM-package}
\alias{hSDM}
\docType{package}

\title{hierarchical Bayesian species distribution models}

\description{\code{hSDM} is an R package for estimating parameters of
  hierarchical Bayesian species distribution models. Such models allow
  interpreting the observations (occurrence and abundance of a species)
  as a result of several hierarchical processes including ecological
  processes (habitat suitability, spatial dependence and anthropogenic
  disturbance) and observation processes (species
  detectability). Hierarchical species distribution models are essential
  for accurately characterizing the environmental response of species,
  predicting their probability of occurrence, and assessing uncertainty
  in the model results.}

\details{
  \tabular{ll}{
    Package: \tab hSDM\cr
    Type: \tab Package\cr
    Version: \tab 1.0\cr
    Date: \tab 2012-11-19\cr
    License: \tab GPL-3 \cr
    LazyLoad: \tab yes\cr
  }
}

\author{
  \tabular{l}{
    Ghislain Vieilledent <ghislain.vieilledent@cirad.fr>\cr
    Andrew M. Latimer <amlatimer@ucdavis.edu>\cr
    Alan E. Gelfand <alan@stat.duke.edu>\cr
    Cory Merow <cory.merow@gmail.com>\cr
    Adam M. Wilson <adam.wilson@yale.edu>\cr
    Frederic Mortier <frederic.mortier@cirad.fr>\cr
    John A. Silander Jr. <john.silander@uconn.edu>\cr
  }
}

\keyword{biodiversity}
\keyword{species distribution models}
\keyword{hierarchical Bayesian models}
\keyword{spatial correlation}
\keyword{intrinsic CAR model}
\keyword{conditional autoregressive model}
\keyword{MCMC}
\keyword{Markov Chains Monte Carlo}
\keyword{Metropolis algorithm}

