\name{neighbors.Latimer2006}
\alias{neighbors.Latimer2006}

\docType{data}

\title{Neighborhood data (from Latimer et al. 2006)}

\description{Data come from a small region including 476 one minute by
one minute grid cells. This region is is a small corner of South
Africa's Cape Floristic Region, and includes very high plant species
diversity and a World Biosphere Reserve. The data frame can be used as
an example for several functions in the \code{hSDM} package.}

\format{
  
  \code{neighbors.Latimer2006} is a vector of 3542 integers indicating
  the neighbors (adjacent cells) of each spatial cell. The vector is of
  the form c(neighbors of cell 1, neighbors of cell 2, ... , neighbors
  of the last cell).
  
}

\source{Latimer et al. (2006) \emph{Ecological Applications}, Appendix B}

\references{

  Latimer, A. M.; Wu, S. S.; Gelfand, A. E. and Silander, J. A. (2006)
  Building statistical models to analyze species
  distributions. \emph{Ecological Applications}, 16, 33-50.

}

\keyword{datasets}
