\name{punc10}

\alias{punc10}

\docType{data}

\title{Occurrence data for \emph{Protea punctata} Meisn. in the Cap
Floristic Region}

\description{The species data were collected by the Protea Atlas Project
of South Africa’s National Botanical Institute.}

\usage{punc10}

\format{
  
  \code{cfr.env} is a data frame with 2934 presence-absence observation points.
  \describe{
    \item{\code{Occurrence}}{presence (1) or absence (0) of the species}
    \item{\code{lon}}{longitude}
    \item{\code{lat}}{latitude}
  }
  
}

\source{Cory Merow's personal data}

\references{

  Latimer, A. M.; Wu, S. S.; Gelfand, A. E. and Silander, J. A. (2006)
  Building statistical models to analyze species
  distributions. \emph{Ecological Applications}, 16, 33-50.

}

\keyword{datasets}
